requirements.txt
streamlit>=1.22.0
opencv-python-headless>=4.5.3,<4.6.0
Pillow>=8.2.0
imutils>=0.5.4
numpy>=1.21.0
requests>=2.26.0
# EasyOCR dependencies
torch>=1.7.0,<2.0.0
torchvision>=0.8.1,<0.15.0
easyocr>=1.4.1